XmlColumnEditPanel now has its own file. Is now public class
